// partner-edit/frontend/src/app/api/model/partner-edit-dto.ts
export interface PartnerEditDto {
  partnerNumber?: number;
  alphaCode?: string;
  name1?: string;
  name2?: string;
  name3?: string;
  firstname?: string;
  street?: string;
  houseNumber?: string;
  postalCode?: string;
  city?: string;
  type?: string;
  groupType?: string;
  groupNumber?: number;
}